move simulations to the same file, and up a directory to clean up the structure of the module
